Fixes dependencies with the newer versions of Devise and Capistrano.
database_acts_as_authenticatable is depricated with the newest version, I changed the migration file to the newer style so it is now current.

initializers/devise.rb: added secret key

.ruby-version: I use rbenv and had to delete the ruby version to get Passkeep to work on my machine. If RVM requires use of this file just add to the top of the file without quotes: "ruby-2.0.0"

database.yml was necessary for use with postgres instead of mysql

capistrano-colors was crashing the capistrano. Hopefully the terminal is legible and the upgrade is of an appropriate benefit.